+++ packages/grafana-data/src/types/dataLink.ts
@@ -1,0 +1,5 @@
+export interface DataLink {
+  url: string;
+  title: string;
+  targetBlank?: boolean;
+}
