+++ packages/grafana-data/src/types/index.ts
@@ -1,5 +1,8 @@
 export * from './data';
+export * from './dataLink';
 export * from './logs';
 export * from './navModel';
 export * from './time';
+export * from './threshold';
 export * from './utils';
+export * from './valueMapping';
