+++ packages/grafana-ui/src/types/index.ts
@@ -4,6 +4,5 @@ export * from './app';
 export * from './datasource';
 export * from './theme';
 export * from './graph';
-export * from './threshold';
 export * from './input';
 export * from './displayValue';
