+++ packages/grafana-ui/src/types/panel.ts
@@ -123,35 +123,6 @@ export interface PanelMenuItem {
   subMenu?: PanelMenuItem[];
 }
 
-export enum MappingType {
-  ValueToText = 1,
-  RangeToText = 2,
-}
-
-interface BaseMap {
-  id: number;
-  operator: string;
-  text: string;
-  type: MappingType;
-}
-
-export type ValueMapping = ValueMap | RangeMap;
-
-export interface ValueMap extends BaseMap {
-  value: string;
-}
-
-export interface RangeMap extends BaseMap {
-  from: string;
-  to: string;
-}
-
-export interface DataLink {
-  url: string;
-  title: string;
-  targetBlank?: boolean;
-}
-
 export enum VizOrientation {
   Auto = 'auto',
   Vertical = 'vertical',
