+++ packages/grafana-ui/src/utils/displayValue.ts
@@ -3,20 +3,11 @@ import _ from 'lodash';
 
 // Utils
 import { getValueFormat } from './valueFormats/valueFormats';
-import { getMappedValue } from './valueMappings';
 import { getColorFromHexRgbOrName } from './namedColorsPalette';
 
 // Types
-import {
-  Threshold,
-  ValueMapping,
-  DecimalInfo,
-  DisplayValue,
-  GrafanaTheme,
-  GrafanaThemeType,
-  DecimalCount,
-} from '../types';
-import { DateTime, dateTime, Field } from '@grafana/data';
+import { DecimalInfo, DisplayValue, GrafanaTheme, GrafanaThemeType, DecimalCount } from '../types';
+import { DateTime, dateTime, Threshold, ValueMapping, getMappedValue, Field } from '@grafana/data';
 
 export type DisplayProcessor = (value: any) => DisplayValue;
