+++ packages/grafana-ui/src/utils/index.ts
@@ -1,11 +1,9 @@
 export * from './valueFormats/valueFormats';
 export * from './colors';
 export * from './namedColorsPalette';
-export * from './thresholds';
 export * from './displayValue';
 export * from './fieldDisplay';
 export * from './deprecationWarning';
-export { getMappedValue } from './valueMappings';
 export * from './validate';
 export { getFlotPairs } from './flotPairs';
 export * from './slate';
