+++ public/app/features/dashboard/dashgrid/PanelHeader/PanelHeader.tsx
@@ -9,7 +9,8 @@ import templateSrv from 'app/features/templating/template_srv';
 
 import { DashboardModel } from 'app/features/dashboard/state/DashboardModel';
 import { PanelModel } from 'app/features/dashboard/state/PanelModel';
-import { ClickOutsideWrapper, DataLink } from '@grafana/ui';
+import { ClickOutsideWrapper } from '@grafana/ui';
+import { DataLink } from '@grafana/data';
 
 export interface Props {
   panel: PanelModel;
