+++ public/app/features/panel/panellinks/link_srv.ts
@@ -3,9 +3,8 @@ import { TimeSrv } from 'app/features/dashboard/services/TimeSrv';
 import templateSrv, { TemplateSrv } from 'app/features/templating/template_srv';
 import coreModule from 'app/core/core_module';
 import { appendQueryToUrl, toUrlParams } from 'app/core/utils/url';
-import { DataLink, VariableSuggestion, KeyValue, ScopedVars } from '@grafana/ui';
-import { TimeSeriesValue, DateTime, dateTime } from '@grafana/data';
-import { deprecationWarning, VariableOrigin } from '@grafana/ui';
+import { VariableSuggestion, KeyValue, ScopedVars, deprecationWarning, VariableOrigin } from '@grafana/ui';
+import { TimeSeriesValue, DateTime, dateTime, DataLink } from '@grafana/data';
 
 export const DataLinkBuiltInVars = {
   keepTime: '__url_time_range',
