+++ public/app/plugins/panel/graph/module.ts
@@ -11,8 +11,8 @@ import { DataProcessor } from './data_processor';
 import { axesEditorComponent } from './axes_editor';
 import config from 'app/core/config';
 import TimeSeries from 'app/core/time_series2';
-import { DataFrame } from '@grafana/data';
-import { getColorFromHexRgbOrName, LegacyResponseData, DataLink, VariableSuggestion } from '@grafana/ui';
+import { DataFrame, DataLink } from '@grafana/data';
+import { getColorFromHexRgbOrName, LegacyResponseData, VariableSuggestion } from '@grafana/ui';
 import { getProcessedDataFrame } from 'app/features/dashboard/state/PanelQueryState';
 import { PanelQueryRunnerFormat } from 'app/features/dashboard/state/PanelQueryRunner';
 import { GraphContextMenuCtrl } from './GraphContextMenuCtrl';
